[ci] Check APK sizes on Azure Pipelines
This target was overlooked in the initial split up and porting efforts
for our APK tests. This target calculates the sizes of or test APKs and
compares them against a baseline. This change ensures that we will
report a failure in our Azure Pipelines build if any project's APK size
drifts too far from the baseline.