[build] fix xabuild on Linux (#4129)
After a recent update of `msbuild` on Linux I started to see this
error when attempting to build our tests (with `make all-tests`):

	"Xamarin.Android-Tests.sln" (default target) (1) ->
	"tests/CodeGen-MkBundle/Xamarin.Android.MakeBundle-UnitTests/Xamarin.Android.MakeBundle-UnitTests.csproj" (default target) (21) ->
	  BuildTests.cs(150,22): error CS0012: The type 'ValueType' is defined in an assembly that is not referenced. You must add a reference to assembly 'netstandard, Version=2.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51'.

Which, after investigation by Jo Shields, turned out to be caused by
`xabuild` not creating a bunch of symlinks to Mono API façades (the
`*-api` directories).  Turns out the symlinks were created only on
macOS.

Fix the problem by enabling the facade symlink creation on Linux.