[docs] Add intended audience for prepare-external-dependencies
Since the prepare external dependencies step is only for proprietary
dependencies, update the wording of the build instructions to let
community contributors know that they should skip it.

`make prepare-external-dependencies` uses the
`PrepareExternalGitDependencies` step in `xaprepare`, which currently
only prepares the monodroid external dependency.

The Mono dependency listed in `.external` is prepared via a separate
`DownloadMonoArchive` step in `xaprepare`.  That step is part of `make
prepare` and does not require running `make
prepare-external-dependencies`.