[Build] Bump Xamarin.LibZipSharp to 1.0.7
Fixes #4100

The version of libzip.so we shipped for linux in
1.0.6 had a dependency on libcrypto.so.1.1. Some
linux based systems only ship with libcrypto.so.1.0.
As a result Xamarin.Android would not build and we give
the following error.

	error MSB4018: System.DllNotFoundException: lib64/libzip.so assembly

Verison 1.0.7 if Xamarin.LibZipSharp fixes this issue by
statically linking libcrypto (and other dependencies) into the
dynamic library. This should allow Xamarin.Android to build and
run on those systems which does not include libcrypto or have
an older version.