[vscode] Fix the command used to run the NUnit Tests
When trying to run the 'Run Xamarin.Android Build Tasks Unit Tests'
command in VSCode we get the following error now

	error MSB4057: The target "RunNUnitTests" does not exist in the project.

This is because the `RunTest.targets` file is now only imported
when you run against the `Xamarin.Android.sln`. So we need
to update this command to use the correct command line.