C++ code cleanup based on extra warnings
Add a number of warnings to the array of arguments passed to the C++
compilers in order to detect suboptimal or incorrect code:

  * `-Wduplicated-branches`
    Detect conditional branches which yield identical code.
  * `-Wduplicated-cond`
	  Detect duplicated conditions in the if-else-then statements
  * `-Werror=return-type`
	  Make `-Wreturn-type` an error. Prevents compiling code with
		functions which require a return value but do not return from all
		code paths.
  * `-Wextra`
	  A bunch of "extra" warnings (too many to list here) which should
		help make the code cleaner and more correct
  * `-Wmisleading-indentation`
	  Warn about code which has indentation that can mislead the developer
		into thinking that a statement following the `if` statement is part
		of the statement's `true` block while in reality it isn't.
  * `-Wnull-dereference`
	  Make the compiler detect potential `null` pointer dereference (it
		can yield some annoying false positives, but it's worth dealing with
		them overall)
  * `-Wpointer-arith`
	  Warn about  that depends on the "size of" a function type or of
		'void'.
  * `-Wshadow`
	  Warn about variables which shadow other variables in outer scope.
		This is a potential source of obscure bugs or confusion and should
		be avoided.

Enabling `-Wextra` also enables warnings about unused function
parameters, thus this commit also adds a bunch of `[[maybe_unused]]`
attributes to a few functions throughout the code. Some of those
parameters are used only in certain contexts (e.g. only during Android
or Windows builds) and thus they'll generate the warning in other
contexts.

In addition to the above, clean up the API a bit by removing actually
not used parameters from a handful of functions.