[Xamarin.Android.Build.Task] use global:: for GeneratedCodeAttribute
Fixes http://work.azdo.io/1057222

We had the global:: namespace prefix on ALL namespaces
in the Resource.Designer.cs file EXCEPT GeneratedCodeAttribute :facepalm.
So users can get errors like

	The type or namespace name 'CodeDom' does not exist in the namespace Foo.Bar.System

If they use `System` in their defualt namespace for the project.
This commit adds it, so users who use `System` in their
assembly namespaces will not be hit by this in the future.