[Xamarin.Android.Build.Tasks] support for $(AdbTarget) (#4166)
Fixes: https://developercommunity.visualstudio.com/content/problem/854863/unable-to-deploy-to-physical-device-when-emulator.html
Fixes: http://work.azdo.io/1056857

We got some reports that our App Bundle MSBuild tasks fail when you
have multiple devices/emulators connected to adb.

We were not using the `$(AdbTarget)` MSBuild property at all, which is
actually a slightly different parameter for `bundletool`:

    --device-id: (Optional) Device serial name. If absent, this uses the
        ANDROID_SERIAL environment variable. Either this flag or the environment
        variable is required when more than one device or emulator is connected.
        Used only if connected-device flag is set.

It appears that only a value such as `-s emulator-5554` would work for
the `--device-id` switch. To make things compatible, I think we should
just look for a space character and use the text as-is after it.

I added some tests that validate the generated command-line for
`bundletool`-related MSBuild tasks.