[ci] Check APK sizes on Azure Pipelines (#4133)
Future-Followup: https://github.com/xamarin/xamarin-android/issues/4169

The `RecordApkSizes` target was overlooked in the initial split up
and porting efforts for our `.apk` tests (fd8336c1).  `RecordApkSizes`
calculates the sizes of or test `.apk` files and compares them against
a baseline.  This change ensures that we will report a failure in our
Azure Pipelines build if any project's `.apk` size drifts too far from
the baseline.

In order for the `.apk` size comparisons to continue to "make sense" --
in that many of the sizes in
`TestResult-Xamarin.Forms_Tests-values-Release.csv` are for AOT builds,
while fd8336c1 didn't build those projects with AOT enabled, resulting
in incomparable results -- we need to ensure that we're building the
appropriate build config, by setting `$(AotAssemblies)`=True.  This
fixes a build-time failure:

	error MSB4018: The "ProcessApkSizes" task failed unexpectedly. [/Users/runner/runners/2.164.3/work/1/s/tests/Xamarin.Forms-Performance-Integration/Droid/Xamarin.Forms.Performance.Integration.Droid.csproj]
	error MSB4018: System.Collections.Generic.KeyNotFoundException: The given key 'apk-Release-Profiled' was not present in the dictionary. [/Users/runner/runners/2.164.3/work/1/s/tests/Xamarin.Forms-Performance-Integration/Droid/Xamarin.Forms.Performance.Integration.Droid.csproj]

Finally, update `TestResult-Xamarin.Forms_Tests-values-Release.csv`
to contain values which won't cause the build to fail.  Unfortunately
these sizes are larger, and we're not entirely sure why [^0].
Issue #4169 is tracking that investigation.

[^0]: We have a *guess*: xamarin/java.interop@3b24a2c added
      `[MethodImpl (MethodImplOptions.AggressiveInlining)]` to
      multiple methods, which *presumably* increases AOT output.
      This might well be an acceptable size increase.