[Xamarin.Android.Build.Tasks] Move Generator.cs error strings into .resx file
Context: [0342fe56][0]
Context: https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1009374/

Move the error strings for `Generator.cs` into the `.resx` file.

Other changes:

Change "JavaTypeInfo" to "Java type" in the messages because the old
class name `JavaTypeInfo` no longer exists.

Change "create" to "generate" in the messages to match the name of the
failing method and hopefully reduce potential confusion with
*instantiating* Java types.

Correct a "The the" typo in the comment for one of the existing
resources in the `.resx` file.

[0]: https://github.com/xamarin/xamarin-android/commit/0342fe5698b86e21e36c924732ff135b9a87e4af