[Xamarin.Android.Build.Tasks] fixes for jnimarshalmethod-gen (#4181)
Context: https://github.com/xamarin/xamarin-android/pull/4171

When working on PR #4171, I was hitting an NRE on this line:

    marshalTypes.Add (GetType ("Mono.Android", "Java.Interop.TypeManager/JavaTypeManager/__<$>_jni_marshal_methods"));

Somehow `Mono.Android.dll` was missing?

    Java.Interop.TypeManager/JavaTypeManager/__<$>_jni_marshal_methods

After more investigation, I came upon this `<ItemGroup>` that seemed
wrong:

    <ItemGroup Condition=" '$(AndroidGenerateJniMarshalMethods)' == 'True' ">
      <_JniAssembliesToLink Include="@(ResolvedAssemblies->'$(_JniMarshalMethodsOutputDir)%(Filename)%(Extension)')" />
      <_AssembliesToLink Condition="Exists(%(Identity))" Include="@(_JniAssembliesToLink)" />
      <_AssembliesToLink Condition="!Exists(@(JniAssembliesToLink))" Include="@(ResolvedAssemblies)" />
    </ItemGroup>

In this example the `Condition` on the second `@(_AssembliesToLink)`
would always evaluate to `true`. Thus `Mono.Android.dll` was in the
list twice and PR #4171 happened to cause this bug, because the
assemblies end up in a different order.

The fix is to more appropriately merge the item groups here using item
metadata.

I also made the linker throw a better exception message instead of an
NRE if something like this happens in the future.