[Xamarin.Android.Build.Tasks] fix potential AsyncTask.Log calls (#4187) (#4205)
We have had a continuing concern of accidentally calling methods on
`AsyncTask.Log` from a background thread.  If it happens, this can
cause a hang inside Visual Studio on Windows.

I finally had the idea of creating a build error if we inadvertently
do this:

	[Obsolete("You should not use the 'Log' property directly for AsyncTask. Use the 'Log*' methods instead.", error: true)]
	public new TaskLoggingHelper Log { get; }

This uncovered 26 build errors!  Some were actually on the main
thread and completely OK, but several were not OK.

In most of the places, I could simply change `Log.LogDebugMessage()`
to `LogDebugMessage()`.

However a few of the more troublesome places:

  * `ManifestDocument` required a `TaskLoggingHelper` for its ctor.
    I moved this to only the methods that need it.

  * `NdkUtils` and `NdkUtilsOld` I used `Action<string>` and
    `Action<string, string>` callbacks instead.

  * The `<Aot/>` task had an empty `ValidateAotConfiguration()`
    method I removed.

Going forward, it should be a lot harder for us to introduce a hang
by using `AsyncTask.Log`.