[Xamarin.Android.Build.Tasks] CleanBOM could leave behind a temp file (#4214)
Fixes: http://work.azdo.io/774576

In reviewing `Path.GetTemp*` usage, I found the
`MonoAndroidHelper.CleanBOM` method could leave behind temp files if
an exception is thrown. It did not use a `try-finally` block.

I took the time to refactor the method and add a general test for it.

The other places I reviewed seem to be OK. They have `File.Delete`
calls in a `try-finally` block. Some are at the MSBuild task level.