[Xamarin.Android.Build.Tasks] Re-enable Resources.Designer.cs generation
Now that [03743f32][0] has switched the
`Xamarin.Android.Build.Tasks.csproj` file to the short-form project
style, some item metadata needs to be added explicitly to
`Resources.resx` and `Resources.Designer.cs` to tell Visual Studio to
generate `Resources.Designer.cs` from `Resources.resx` as expected.  The
short-form project style does add the files automatically to the
appropriate `@(EmbeddedResource)` and `@(Compile)` item lists via the
default wildcard includes, but it doesn't add this metadata.

The new items are almost the same as in [0342fe56][1], but they use the
`Update` attribute instead of the `Include` attribute because now the
items are already included via the wildcard includes.

[0]: https://github.com/xamarin/xamarin-android/commit/03743f32aec3b02edc280886061b12e0dde0bd1c
[1]: https://github.com/xamarin/xamarin-android/commit/0342fe5698b86e21e36c924732ff135b9a87e4af