[Xamarin.Android.Build.Tasks] Dont always lowercase "*Layout" element values
Fixes http://work.azdo.io/1042658

Commit a6387737 made the assumption that ALL `*Layout` view
extension values needed to be lowercased. This is not always
the case. In the Issue listed above the user defined custom
enumeration values which were CamelCase. These were not lowercased
by our build process. So when we lowercased the value in the
layout we got the following build error

	Error		‘fixedsize’ is incompatible with attribute ChildLayout (attr) enum [Auto=0, FixedHeight=2, FixedSize=3, FixedWidth=1, MaxHeightAlignBottom=16, MaxHeightCenter=32, MaxWidthAlignRight=4, MaxWidthCenter=8]

To work around this issue this PR logs all of the `attr` values it
comes across during the conversion of the casing. If the value
for the view extension exists in that list it will not be lowercased.
Otherwise the code will proceed as normal.

The order in which the xml files are processed also changed to make
this work. Before we just processed them in the order they were found.
Now we process them by the directory they are in and in reverse. This
is so the `values` folder which contain the `attribs.xml` based files
are processed first.