[size] Update apk size reference
Mono.Android increase comes from these commits

https://github.com/xamarin/xamarin-android/commit/476921a7e526ffd522245de2bcfef93f1a91fd0a
https://github.com/xamarin/xamarin-android/commit/1273970bcc77191177617225d21ac9772afd62cd

First is NDK bump, which adds about 40-50kbytes to our tests apks and
we cannot do much about it.

Second is adding code to the test, so the size increase is
expected. In future we plan to stop using Mono.Android test for size
regression testing as it changes frequently. We will be using a
HelloWorld sample, which will be updated, so that it is close to XA
app template.

Then there is XF test size increase coming from Java.Interop bump
https://github.com/xamarin/xamarin-android/commit/5945e9a77988a01ceb062dcc82bda6c6bf407e14

For this one I have opened an issue https://github.com/xamarin/xamarin-android/issues/4231 to analyze the size increase from enabling NRT.