[Xamarin.Android.Build.Tasks] deprecate [assembly: DoNotPackage]
The `<ResolveAssemblies/>` MSBuild task currently looks for an
attribute in every assembly:

    [assembly: Java.Interop.DoNotPackage ("foo.jar")]

`foo.jar` would be used at compile time for java code, but not
included in the final Android app. This seems like the wrong way to go
about things, you can simply add to the `@(AndroidExternalJavaLibrary)`
item group for the same behavior.

I also found very little usage of this on Github:

    https://github.com/search?l=C%23&p=1&q=%5Bassembly%3A+DoNotPackage&type=Code

Most of the projects using this are unsupported libraries from the
Xamarin Component store.

If we plan to completely remove `<ResolveAssemblies/>` this is a piece
we will have to deprecate for a period of time first.

Co-Authored-By: Brendan Zagaeski <brzaga@microsoft.com>